fix #3
MultiscaleRun was heavily dependent on bb5. Its loss made the program fail in various parts.
Here I want to start from the base and rebuild from the ground up checking with the `tiny_CI` circuit.
The only circuit I was able to save for posterity.

It is too slow to run on a single core. We will employ mpirun

- remove bloodflow dependencies
- remove steps dependencies
-  fix spack installation problems
- leave julia for now, remove diffeqpy as it is problematic
- the latest release of `neurodamus-models` gives a seg-fault for glu2 and cadifus.
  We should use develop
- add `setup.py` to set spack, python and prepare the tiny_CI simulation
- remove ci. I will reactivate once something works. It is less important now
- update readme
- add the testing in the readme